refactor : unify db connect function
